Add Distance trait. Mv normalized distance to distance
